requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
suds-jurko==0.6
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
